src/requirements.txt
requests[security] == 2.31.0
python-dotenv == 1.0.1
mysql-connector-python == 8.3.0
DateTime == 5.4